Add Node and Graph classes to manage graph structure
Adds the Node and Graph classes to manage the structure of a graph. The Node class represents a node in the graph, with properties such as ID, coordinates, floor, building, node type, long name, short name, and edges. The Graph class represents the entire graph and provides methods to add nodes and edges, as well as retrieve nodes by ID. An example usage is included at the end of the file.